+++ src/components/IncidentRow.tsx
@@ -2,7 +2,7 @@ import { Component } from 'solid-js'
 import { useNavigate } from 'solid-app-router'
 import { format } from 'date-fns'
 
-import { Incident } from '../types/Incident'
+import { Incident } from '../types/incident'
 import { getIncidentSeverityIcon } from './IncidentSeverityIcons'
 
 type Props = {
