+++ src/pages/incidents/assigned.tsx
@@ -3,7 +3,7 @@ import { Component, For } from 'solid-js'
 import IncidentsSidebar from '../../components/IncidentsSidebar'
 import IncidentRow from '../../components/IncidentRow'
 import AppLayout from '../../components/layouts/AppLayout'
-import { Incident, IncidentSeverity } from '../../types/Incident'
+import { Incident, IncidentSeverity } from '../../types/incident'
 import IncidentsTable from '../../components/IncidentsTable'
 
 const INCIDENTS = [
