+++ src/types/ui.ts
@@ -2,3 +2,8 @@ export type IncidentsSidebarItem = {
   title: string
   route: string
 }
+
+export type DropdownOption = {
+  id: string
+  label: string
+}
