+++ src/utils/getOption.ts
@@ -1,4 +1,4 @@
-import { DropdownPropertyOption } from '../types/Incident'
+import { DropdownPropertyOption } from '../types/incident'
 
 export const getOption = (id: string, options: DropdownPropertyOption[]) =>
   options.find((option) => option.id === id)
